Initial SoC implementation - Lattice XP2-5 (bravia development kit)
This is the initial implementation that's fully functional. When the
bitstream is loaded the CPU is held in reset, it stays in reset until
a program is loaded into RAM over the SPI port.

There are two wishbone masters, the CPU and a SPI debug port. They are
connected to the system bus through a priority arbiter giving
preference to the CPU interface. The SPI port is initially used to
load a program into the block ram for the CPU to execute. It's also
useful to dump memory while the CPU is running for debug. The runtime
use case for this SPI port is to provide a shared memory style
interface between code running on a host system and code running on
the SoC. The program used on the host (raspberry pi in this case)
uses the spidev device driver. This program is called 'robotsoc-io'
and source code is in the tools directory.

The FPGA is loaded via JTAG using a SVF file, created using the
make-svf.sh script. This script makes a SVF for programming to flash
or loading to RAM. The 'svfload' source code is here.

  https://github.com/northwoodlogic/pi-industrial-robottools

A sample LED blinky program is located in the sw directory. This demo
may be compiled using a risc-v toolchain configured for baremetal
rv32i. Follow these instructions to setup the toolchain.

  https://github.com/riscv-collab/riscv-gnu-toolchain

  ./configure --prefix=/opt/riscv --with-abi=ilp32 --with-arch=rv32i
  make

Load program as follows:

    # ./load-bin.sh hello.bin
    write: 0x100000=0x00000001
    Loading mem file: hello.bin
    mem file, read 16384 bytes
    write: 0x100000=0x00000004

Watch blinky lights. There are no I/O peripherals except LEDs.