Implement 2x R/C receiver pulse position inputs
Each receiver channel is sampled into an 8-bit value with an
additional status bit indicating if receiver lock is established or
not. If the R/C receiver stops outputting a signal, or the PPM signal
gets stuck high or low then the channel value resets to mid-position
and the 'receiver locked' bit is cleared. If the measured pulse width
is outside a 1-2mS window it is considered invalid and receiver lock
is lost.