Implement 8x R/C transmitter pulse position outputs
This implements 8x PPM signals that can be plugged directly into a
standard R/C servo or speed controller. It also implements a serial
PPM stream output that could be fed into a RF or IR transmitter enable
for wireless operation, or fed into a demuxer (decade counter) for
off-board servo channel control.

There is a 168mS watchdog timeout. If it expires all 8x holding
registers are reset to mid-position. The module reset is also tied in
with the emergency brake input. Therefore, E-brake assertion also
resets to mid-position. The watchdog is reset when any 1 of 8 holding
registers are written.