Adjust memory map, add more bram
Memory is increased by 2KB, adding the 9'th block RAM into the main
system memory. The linker script and loader program are adjusted
accordingly.

The I/O peripheral address map is shuffled around. The new layout is
documented in the test program. In addition, the test program is
reworked with an example how to create a software interval timer using
the free running 1mS ticker.