Add a 2nd CPU
The bus arbiter is modified to support another bus master. The arbiter
includes another output indicating which bus master currently owns the
bus. This is used as the HART ID and made available in the previously
unused status register byte at address 0x400002. The ID is a one-hot
encoded value, id[0] == CPU0, id[1] == CPU1. id[2] == SPI.

The program loader is multi-cpu aware, able to load and release
0/1 or both CPUs from reset.

A test blinky program is added. It's written in assembly and toggles a
LED index matching the HART id.

The crt0.s start does not yet support SMP mode as it doesn't configure
separate stack pointers for each CPU. Therefore, programs linked with
this startup code may execute on either CPU but not both at the same
time.