Implement SRAM controller
This module implements a 32-bit wishbone bus to 8-bit async SRAM
(131072 bytes) controller with individual byte enables. This also
includes a test bench and notes about running a minimal bitstream on
the FPGA for testing memory accesses using a SPI interface of an
external device.

The constraints file is updated to place SRAM interface signals at the
correct pads, but this controller is not yet integrated into the
design. The result is warnings are emitted during synthesis but is
otherwise harmless.

Finally, add datasheets for the SRAM part and XP2 Brevia development
board.